Alterações

+++ src/modules/medicos/medicoEspecialidade/services/DeleteMedicoEspecialidadeService.ts
@@ -1,5 +1,6 @@
 import { getCustomRepository } from 'typeorm';
 import MedicoEspecialidadeRepository from '../typeorm/repositories/MedicoEspecialidadeRepository';
+import AppError from '@shared/errors/AppError';
 
 interface IRequest {
   idMedicoEspecialidade: number;
@@ -8,6 +9,13 @@ interface IRequest {
 export default class DeleteMedicoEspecialidadeService {
   public async execute({ idMedicoEspecialidade }: IRequest): Promise<void> {
     const medicoEspecialidadeRepository = getCustomRepository(MedicoEspecialidadeRepository);
+
+
+    const medicoEspecialidade = await medicoEspecialidadeRepository.findByMedicoAndEspecialidade(idMedicoEspecialidade);
+
+    if (!medicoEspecialidade) {
+      throw new AppError('MedicoEspecialidade not found.');
+    }
 
     await medicoEspecialidadeRepository.delete(idMedicoEspecialidade);
   }
